+++ lib/chat_gpt_sdk.dart
@@ -9,3 +9,5 @@ export 'src/model/gen_image/request/generate_image.dart';
 export 'src/model/gen_image/response/GenImgResponse.dart';
 export 'src/model/openai_model/openai_models.dart';
 export 'src/model/openai_engine/engine_model.dart';
+export 'src/model/chat_complete/request/ChatCompleteText.dart';
+export 'src/model/chat_complete/response/ChatCTResponse.dart';