+++ lib/src/model/chat_complete/response/message.dart
@@ -1,0 +1,16 @@
+class Message {
+  final String role;
+  final String content;
+
+  Message({required this.role, required this.content});
+
+  factory Message.fromJson(Map<String, dynamic> json) => Message(
+        role: json["role"],
+        content: json["content"],
+      );
+
+  Map<String, dynamic> toJson() => {
+        "role": role,
+        "content": content,
+      };
+}